requirements-dev.txt
black
catboost
cython
dill
libsvm
lightgbm
mleap
numpy
openpyxl
pandas
pyspark
pytest
pytest-cov
pytest-spark
ruff
scikit-learn>=1.2.0
scipy
wheel
xgboost==1.7.5
onnxruntime
